continue rework merge from master, get backpack implementation in place

+++ TreasurePleasure/app/src/main/java/treasure/pleasure/model/UpgradeCenter.java
@@ -10,7 +10,7 @@ public class UpgradeCenter {
         this.belongsTo = p;
     }
 
-    int upgradeDropBonus(Item i) {
+    double upgradeDropBonus(Item i) {
        return belongsTo.getDropBonus() + 1;
     }
